Change DTWAIN_SetTwainLog() first parameter to a DWORD

+++ source/cpp/ctldtwaininit.cpp
@@ -999,7 +999,7 @@ void LoadStaticData(CTL_TwainDLLHandle* pHandle)
     }
 }
 
-DTWAIN_BOOL DLLENTRY_DEF DTWAIN_SetTwainLog(LONG LogFlags, LPCTSTR lpszLogFile)
+DTWAIN_BOOL DLLENTRY_DEF DTWAIN_SetTwainLog(DWORD LogFlags, LPCTSTR lpszLogFile)
 {
     LOG_FUNC_ENTRY_PARAMS((LogFlags, lpszLogFile))
     auto [pHandle, pSource] = VerifyHandles(nullptr, DTWAIN_VERIFY_DLLHANDLE);
