+++ source/h/ctllogcalls.h
@@ -23,7 +23,7 @@
 
 namespace dynarithmic
 {
-    long& GetLogFilterFlags();
+    uint32_t& GetLogFilterFlags();
     std::string CTL_LogFunctionCallHelper(const char* pFuncName, int nWhich, const char* pOptionalString = nullptr);
     std::string CTL_LogFunctionCallA(int32_t logFlags, const char *pFuncName, int nWhich, const char* pOptionalString = nullptr);
 }
